Indecies should be integers ;)

+++ app/views.py
@@ -40,10 +40,10 @@ def search(search_string=None):
                 stats['median_price'] = median(prices)
                 stats['std_deviation'] = std(prices)
                 stats['worst_price'] = min(prices)
-                stats['bad_price'] = prices[floor(len(prices)*5/6)]
-                stats['poor_price'] = prices[floor(len(prices)*4/6)]
-                stats['good_price'] = prices[floor(len(prices)*2/6)]
-                stats['great_price'] = prices[floor(len(prices)*1/6)]
+                stats['bad_price'] = prices[int(floor(len(prices)*5/6))]
+                stats['poor_price'] = prices[int(floor(len(prices)*4/6))]
+                stats['good_price'] = prices[int(floor(len(prices)*2/6))]
+                stats['great_price'] = prices[int(floor(len(prices)*1/6))]
                 stats['best_price'] = max(prices)
             stats['trend'] = [[mktime(i.close.timetuple()), i.price_per_stick] for i in closed_auctions if i.price_per_stick is not None]
